Fix dashboard 500 error and add time range controls
Change floors_climbed/intensity_minutes to float in schema to match
Garmin's actual data. Add "All Time", "1 year", and custom date range
options across Overview, Sleep, and Body tabs.

+++ app/schemas.py
@@ -55,8 +55,8 @@ class DailySummaryOut(BaseModel):
     max_stress: int | None = None
     body_battery_high: int | None = None
     body_battery_low: int | None = None
-    floors_climbed: int | None = None
-    intensity_minutes: int | None = None
+    floors_climbed: float | None = None
+    intensity_minutes: float | None = None
 
     class Config:
         from_attributes = True
